HAI-1378 Move updateHanke tests to a separate test class
HankeServiceITests has been a bit unwieldy. It's been really long, and
more than half of the tests have been for `updateHanke`. They have been
sprinkled around the test class between blocks of tests for other
methods.

Move the tests for `updateHanke` to a separate test class. This
consolidates them in a single place and makes them easier to manage.
